+++ android/app/src/main/java/com/zzang/chongdae/presentation/view/home/HomeFragment.kt
@@ -13,7 +13,7 @@ import com.zzang.chongdae.data.remote.api.NetworkManager
 import com.zzang.chongdae.data.remote.source.impl.OfferingsDataSourceImpl
 import com.zzang.chongdae.data.repository.remote.OfferingsRepositoryImpl
 import com.zzang.chongdae.databinding.FragmentHomeBinding
-import com.zzang.chongdae.presentation.view.detail.OfferingDetailActivity
+import com.zzang.chongdae.presentation.view.offeringdetail.OfferingDetailActivity
 import com.zzang.chongdae.presentation.view.home.adapter.OfferingAdapter
 
 class HomeFragment : Fragment(), OnArticleClickListener {
