+++ android/app/src/main/java/com/zzang/chongdae/presentation/view/offeringdetail/ParticipationClickListener.kt
@@ -1,4 +1,4 @@
-package com.zzang.chongdae.presentation.view.detail
+package com.zzang.chongdae.presentation.view.offeringdetail
 
 interface ParticipationClickListener {
     fun onClickParticipation()
